refactor(pptx_generator): reorganize Markdown parsing and value objects, consolidate logic into `service` for clarity and maintainability
