+++ lib/pptx_generator/valueobject.py
@@ -1,4 +1,6 @@
 from dataclasses import dataclass
+from typing import Iterable, cast
+from bs4.element import Tag, PageElement
 from lxml import etree
 
 
@@ -105,3 +107,68 @@ class TextContent:
         old = t_elem.text
         t_elem.text = self.text
         return old
+
+
+# ---------------- Markdown Value Objects ----------------
+
+
+@dataclass(frozen=True)
+class HtmlTextExtractor:
+    """BeautifulSoup の Tag から安全にテキストを抽出する値オブジェクト。
+
+    - separator と strip を固定し、型不整合をこの層で閉じ込める。
+    - 外部からは型安全な API のみを露出する。
+    """
+
+    separator: str = " "
+    strip: bool = True
+
+    def extract_all(self, elements: Iterable[Tag]) -> list[str]:
+        results: list[str] = []
+        for e in elements:
+            text = self.extract(e)
+            if text:
+                results.append(text)
+        return results
+
+    def extract(self, element: Tag) -> str:
+        """BeautifulSoup.get_text の型スタブ不整合に関するエクスキューズをここに集約。
+        - いくつかの bs4 の型スタブでは get_text のシグネチャが実装と食い違い、
+        - separator/strip をキーワード指定しても型検査で誤検出される場合がある。
+        - 本実装はランタイムでは正しく動作するため、誤検出のみを抑止する目的で
+        - `# type: ignore[arg-type]` を最小範囲（この呼び出し行）に限定して付与している。
+        - 目的は「VO 内にライブラリ依存と型差異を閉じ込め、外部には型安全な API だけを露出する」こと。
+        """
+        el = cast(PageElement, element)
+        return el.get_text(separator=self.separator, strip=self.strip)  # type: ignore[arg-type]
+
+
+@dataclass(frozen=True)
+class BulletList:
+    """箇条書き1セット（ul/ol）を表す値オブジェクト。"""
+
+    items: list[str]
+
+
+@dataclass(frozen=True)
+class TableRecord:
+    """表の1レコード（1行）を表す値オブジェクト。"""
+
+    cells: list[str]
+
+
+@dataclass(frozen=True)
+class Table:
+    """表全体を表す値オブジェクト（先頭にヘッダ行を含む）。"""
+
+    records: list[TableRecord]
+
+
+@dataclass(frozen=True)
+class MarkdownSection:
+    """Markdown の内容を中立的な構造にした値オブジェクト。"""
+
+    title: str | None
+    paragraphs: list[str]
+    bullet_list: BulletList | None
+    table: Table | None
